Got rid of symbolIndex and single list of functions and symbols
Compiler had a list of functions and symbols. These were supposed to be the list of all functions and symbols in all structs. But how do you know which is which. Now Compiler just has a list of top level Structs and each Struct has a list of local symbols, functions and structs. So searching for a symbol requires you to traverse into each object. Everything compiles again. But we're still not saving the ASTs anywhere.