Added annotation support
Each ASTNode now hold an annotationIndex, which is the current line in the source file when we started parsing code correcsponding to the ASTNode. When doing CodeGen we add the current code address to this line of the annotations. So now the source line is shown with the code generated from it.