Changed Functions getting passed around as pointers to smart pointers
Functions were being allocated out of a vector and then a pointer to that function was being passed around. But when you add more functions the vector gets reallocated and the pointer changes. So now we keep a vector of smart function pointers. A function gets allocated as needed and a smart pointer to it is stored in the vector.