Fixed bug where wrong type was used for NE test
I started passing resultType to OpNode. But this is just the type of the result, used for the next operation. The operation type is the type of the left operand (or right if there's no left). This is true unless we're doing logical operations (LAND, LOR, LNOT), where the operation is done with the uint8 type no matter what (and we cast into uint8 when not).

Float comparisons now work.