Changed how operands are accessed in Interpreter
There used to be a single getOpnd, which was used for both signed and unsigned and was confusing. It also had a "byte" param, which was really more like OpSize, but not always. Changed everything to have getUOpnd and getIOpnd and to take OpSize param. And now OpSize is correct for all cases.