Changed how Interpreter is called
Added size of top level struct to executable. Now when Interp is contstructed, the signature, entry point and size are read and an allocation is done on the stack for top level struct. Then the Y register is set to the start of the top level struct. Then the caller pushes params onto the stack and calls execute. The execute method does everything else as before.