Support for vars in the struct.
Now member vars of the current struct are indexed off of Y. Right now this only works for the top-level struct. We need to handle calling a member function of another struct, pushing the current Y reg and setting it to the self pointer of the called struct member function.