WIP - Started implementing constant support.
Constants are no longer at global scope. They can  be at struct or function scope. They are stored in the executable before functions and are accessed with the IndexMode of Constant. This is an address relative to the start of the constant store, which is in read-only memory. Syntax works and constants are stored, but still need to implement accessing them