Fixed bug negating constant value
A constant gets pushed as what looks like an unsigned value. So if you negate it casting gets all screwed up. So now if a constant is negated it is first cast to the corresponding signed type. This is free. It just changes the type in the AST so all the subsequent casts work correctly.