Got pointers working
Now '&' and '*' operators are working and pointers get assigned and derefed as needed.
Also added type checking so pointers can only be assigned to other pointers of the same type.

There is still an issue where trying to pass a pointer when a scalar is expected doesn't generate an error