Added optimization for simple assignment cases
An assignment like this 'a = 10' used to generate PUSHREF a, PUSH 10, POPDEREF. Now it generates PUSH 10, POP a. To do this I reversed the order of push for assignment. Used to be ref, value. Now it's value, ref. That puts the ref next to the popderef to make it easier to optimize.

I also had to implement the POP opcodes which had never been used before.