Added looping tests.
Now you can pass the -t option and it will do a 1 pass test with 2 params, a uint8 and a uint16. And the -l option will pass 5 params, as though running the flicker command. It will init and then loop 10 times.

PostLightEffects is not working right but it's running the test 10 times