Fixed bug where sometimes return values were not pushed.
There was logic embedded in the Function object that said whether or not the return value from that function should be pushed or dropped. But it was a bool in the Function object itself so if ever there was a call that didn't need the return value that bool would be cleared and no other call to that function would ever push the return value. Now that bool is kept in the FunctionCallNode so every call can have a different value