Fixed a bug in handling short floats
A while back I did an optimization where if a float constant was a small integral value I would emit it as a short integer constant with a CASTI8F after. But in emitting it I changed the value of _i to the integer equivalent of the float. On the second pass this made the float look like a huge number. Fixed it so I don't change the underlying value anymore so it's the same on both passes