Added optimization for scalar constants
Scalar constants were being put in the executable and accessed as constant values using the addressing mode. But it's rarely efficient to do this unless the constants are accessed many times. If that's the case it would be better to store the value in a variable and access it from there. So now scalar constants are treated like literals. This often saves space. For instance storing a value of 1 in an int32 constant takes up 4 bytes and another 2 or 3 bytes to access it. Putting it as a literal constant in the code takes 1 byte.