Changed initialize function to a ctor.
Got rid of the initialize function. Now it is simply the name of the struct, just like in C++.  Every struct has a ctor function, whether it is declared or not and it's always the first function in the struct. like before the function has no name and no return type. One thing this will do is to allow var initialization in the struct to work, but that hasn't been tested yet