Changed ctor to being in the list of functions where is was encountered in the source.
Moving the ctor to be the first function of a struct caused the annotations to get all screwed up and caused the source to not interleave with the generated code properly. So I moved it back. The consequence of this is that we now need another way to locate the ctor. We do this already for the top-level struct while walking through its functions. And we currently don't implement calling the ctor when a struct is instantiated at runtime. When we do we'll have to store a pointer to the ctor function so we can call it during construction.