Added main entry point to executable.
Executable now has a 4 byte mainEntryPoint and a ConstructorEntry point followed by the top-level struct size. The main entry point is simply the function named "main". The intent is to first instantiate the top-level struct on the stack, then call its ctor with the args pushed by the caller. Then the main entry point can be called repeatedly as in the case of the loop() function in Arduino.

The Interpreter side of this is not done yet. But it compiles and the Decompiler properly prints out the new format