Added preliminary support for calling member functions
New MCALL op expects a self pointer on the stack, pointing to the instance of the struct to be called.It will push the old self pointer and replace it with the new one.

Previous self pointer is pushed on the stack below the return address. All the logic expects this so a non-member call pushes a dummy self pointer so everything aligns.

Still have implemented support for member functions. In particular calling the ctor doesn't use the correct self pointer.