Fixed bug related to indexing of arrays of structs and structs containing arrays.
Now ASTNode has a virtual elementSizeInBytes() function. for non-arrays it just gives the size of the node. If it's a struct it gives the size of that struct. For arrays it gives the size of one element. This also gets rid of the use of the passed Compiler pointer in ASTNode for everything except setting the annotationIndex.