Changed how annotations are done
Now to set the annotationIndex (source line number) ASTNode is called from Compiler to set it. And after emitting code a pass is made on the ASTNode tree to set all the code addresses. This gets rid of the need to pass an annotation index to every creation of an ASTNode and hides the annotation logic from every ASTNode subclass. It also gets rid of passing a Compiler pointer to emitCode. Everything is much cleaner. I'm still not happy with the interleaving of the source and executable, but now it will be much easier to fix since it's not so embedded in the code.