WIP - Changing LAND and LOR to elide right hand side
This is part of the C spec. If the lhs of && is false then the rhs is skipped and a false is pushed. Likewise if the lhs of || is true, then the rhs is skipped and a true is pushed.

To do this I changed Op::IF to Op::BRF (branch if the TOS is false), and added a Op::BRT (branch if TOS is true).

Still doesn't work. Branch addresses are wrong