Add support for core.format()
Like print but formats the string into a char buffer.

This also handles "strings" on the stack as well as in ROM. Like in the code, the address of a string in ROM has the stacksize added to it. When greater than stack size, addr has stack size subtracted and then it's accessed in ROM. Otherwise it is accessed from the stack.