Added Clover target
Now the 'clvr' target builds the compiler and installs it in /usr/local/bin/clvr. The clvr-interp target builds the interpreter and installs it in /usr/local/bin/clvr-interp. The Clover target does what the old 'compile' target did. It builds both the compiler and interpreter and allows files to be compiled and tested.