Commented out isPointer for DerefNode again
This is the right thing to do. A deref'ed node is NOT a pointer. Doing this causes TestStructPtr to work because it now considers the deref'ed value to be the underlying type and not a pointer type.

I don't remember why I uncommented it in the first place. I need to be better with comments in the future. For now I'm leaving it uncommented with a comment that says that I think this is right. If I get tempted to uncomment it again, I'll think twice.