Added isIndexable, isAssignable and isPointer to ConditionNode.
These should get their values from the _first node of the condition result (we assume the first and second nodes have the same type). This means that if this if the result of the condition is a String it will get properly upcast when we have the case where int32 is enabled (making the size of varargs 32 bits) but addr size is 16 bits.