WIP - Adding optimizations to 6809 CodeGen
I now keep _lastRegState and _lastPtrState member variables in 6809CodeGen. After every emitCode these are set indicating whether the last value emitted is on the stack or in a register. There is a separate value for Regs and Ptrs because the regs are either D or A and the ptr is X. So for instance you can have a value in D and a value in X so you can store D into 0,X without getting the stack involved at all.

When a emitCode needs an input value (e.g., lhs and rhs values) it looks at _lastRegState. In the case of 2 values on the stack lhs will always be on the stack, but rhs could be in a data reg. So you can avoid popping it. This can be problematic if the lhs value needs to be in the register, like SUB. In this case the rhs value is pushed and the operation is performed with both values on the stack. But for an operator like ADD it doesn't matter which is on the right or left, so it leaves the rhs side in the reg and adds with the TOS. This saves a lot of PSH followed by PUL ops.

The code is not complete. I've currently working on the DotNode. And what I have implemented is not tested at all.