Commented out error checking for RegState
The problem is that _lastRegState really needs to be a stack so we can remember what values got pushed to the stack. But we really don't need to know that because if the wrong thing got pushed to the stack or something is not there when we expect it to be we have far bigger problems to deal with. So really _lastRegState only needs to say what, if anything, is in the A or D reg. I should probably get rid of the RegState::StackI8 and RegState::StackI16.