Implemented optimized case for Switch statement on 6809
I also changed the way it works slightly. Rather than having branch statements in the table, I just have addresses. When the table is searched the matching address is returned in X. If no match X contains the address just past the end of the table where the default statement is expected to be.