Fixed relational, logical and conditional operators
Everything was messed up because I was trying to do the tests backwards (rhs op lhs). So now I emit the rhs first so it is on the stack and lhs is in the reg. So it does normal tests now. I still have to invert the tests because I want it to branch on false. So a < b does of GE test, etc. But all these tests seem to work now. More testing needed