Added branch optimization for if statement and relational operators
When you hit ifState, a branchLabel is set. The relational operator codegen sees this and generates an optimized relational test which simply compares the two values and branches to the branchLabel when false.

There are still other operations to optimize (conditional and logical) but this is a good start. Reduces number of operations from 7 to 3.