Added branch optimization for type cast
If a is an int16 you can do: 'if (b) ...'. In the normal case it will type cast this to 8 bits then compare it to 0. But the cast isn't necessary. If there is a branchLabel on the input to type cast it will simply set the CC according to whether or not the value is 0.