Fixed a bug in std::sort
For some reason all of a sudden the sort function for the clauses in the switch statement is no longer "strict-weak ordering". Maybe this is some added feature of the latest XCode? Anyway, now the CaseClause value method will return the most negative 32 bit value, so it always sorts as the first value