Fixed relOp in the "slow" case
Even in the slow case, where you need to store a 0 or 1 based on whether or not the relational test passed, you can still optimize register loading vs pushing. Now it works in all 3 cases (fully optimized with a branch label passed in and var or const operands), register optimized (no branch label passed in but var or const operands) and unoptimized.