Implemented right and left shift operators
Right now it does a loop in all cases. Several optimizations are possible if rhs is a constant. We can unroll the loop, just clear if shift count is 8/16 or greater, shift left 1 for a right shift of 7, etc. For a 16 bit shift of 8 we can simply copy A to B or vice versa. Lots of things like that are possible.