Add ability to skip return statement without error when a return value is expected.
Some functions that return a value don't need a return statement because all their code paths end with a return. Rather than verifying this, which would be pretty hard, just always add a RET to the end of a function. If a return value is expected, _returnValue will have a value of 0, which will work correctly.