Fixed a bug  where an enum value was not allowed in a binary operation
enums are allowed if both sides are the same enum type. This allows enum values to be used in comparisons, but it also allows them in other operations. So "Foo.A + Foo.B" is legal. Not sure how useful this is, but since enums can be cast to integers it shouldn't be harmful.