Added C preprocessor.
This changes the API to the Compiler. Now you simply instantiate it and it will precompile and compile. For now 2 predefined values are passed to the preprocessor: SUPPORT_FLOAT and SUPPORT_INT32. This will allow tests to be written regardless of the current profile.

This is a full C preprocessor. Eventually I could remove the module support in favor of #include. We'll see.