Fixed a bug where we weren't detecting duplicate identifiers
Clover has function scope. A variable created in an inner block is not hidden from the outer block. So you either need to use different variable names or define the variable in the outer block and then use it in the inner block.

I may eventually implement block scope. But that requires having multiple symbol tables in each function. So it may have to wait a while