WIP - Moved all the module related stuff around to get things to compile on Arduino without having tons of ifdefs
I used to have RUNTIME and COMPILE ifdefs in NativeCore and NativeColor so they could be compiled for the versions that needed to be used during compiling vs running. This was a mess so I got rid of it all and moved the code around so only the runtime stuff got compiled into the Arduino version. The fallout of this is that I had to comment out all the setLight code, So none of that works anymore.

Next I need to add logic for controlling the light hardware into NativeColor (and probably rename it NativeLight since that's what it's really for.