Made runtime module management Arduino friendly
Before I had lots of ifdefs so the Arduino would only build with the module runtimes. I got rid of all that and now have a separate NativeXXXRuntime.cpp file which has all the runtime logic. Also the NativeXXX.h files no longer have a dependency on Module.h so Arduino can build them. Finally, I changed the Module name to be a const char* so Arduino doesn't need to know about std::string. Even though the Runtime doesn't need to know about Module names Arduino would see std::string in the function signature. This way I don't need more ifdefs to deal with that.

The runtime on Mac works and compiles for Arduino. But the Arduino code no longer works. I don't know if that's a problem with the NativeColor module or some other code changes made previously.

The Interpreter no longer needs any knowledge of NativeColor. It's installed from the caller. But the Compiler still has it built-in so I have to fix that.