Changed how custom user functions are called
The method of installing custom functions by using modules was cumbersome and error prone. Now I've added a 'userCall' method to core which passes a unique id and a variable number of parameters. The id has to match between the code and installed user function handler. Right now using for the setLight function for PostLightController.